# api/docs/swagger.yaml
definitions:
  github_com_time_capsule_Api-Gateway-Timecapsule_models.CreateCommentModel:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      memory_id:
        type: string
      user_id:
        type: string
    type: object
...
  github_com_time_capsule_Api-Gateway-Timecapsule_models.PatchCommentModel:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      memory_id:
        type: string
      user_id:
        type: string
    type: object
...
  github_com_time_capsule_Api-Gateway-Timecapsule_models.UpdateCommentModel:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      memory_id:
        type: string
      user_id:
        type: string
    type: object
...
  memory.Comment:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      memory_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  memory.GetAllCommentsResponse:
    properties:
      comments:
        items:
          $ref: '#/definitions/memory.Comment'
        type: array
      count:
        type: integer
    type: object
...
info:
  contact: {}
  description: This is a sample server celler server.
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /v1/comments:
    get:
      consumes:
      - application/json
      description: Get a list of comments.
      parameters:
      - description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      - description: Number of items per page
        in: query
        minimum: 1
        name: limit
        type: integer
      - description: Filter by memory ID
        in: query
        name: memory_id
        type: string
      - description: Filter by user ID
        in: query
        name: user_id
        type: string
      - description: Filter by content
        in: query
        name: content
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/memory.GetAllCommentsResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: List Comments
      tags:
      - Comments
    post:
      consumes:
      - application/json
      description: Create a new comment with the provided details.
      parameters:
      - description: Comment details
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/github_com_time_capsule_Api-Gateway-Timecapsule_models.CreateCommentModel'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new Comment
      tags:
      - Comments
  /v1/comments/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a comment by its ID.
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete Comment
      tags:
      - Comments
    get:
      consumes:
      - application/json
      description: Get details of a comment by its ID.
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/memory.Comment'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get Comment by ID
      tags:
      - Comments
    patch:
      consumes:
      - application/json
      description: Partially update an existing comment.
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      - description: Patched comment data
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/github_com_time_capsule_Api-Gateway-Timecapsule_models.PatchCommentModel'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Partially update Comment
      tags:
      - Comments
    put:
      consumes:
      - application/json
      description: Update an existing comment.
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated comment data
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/github_com_time_capsule_Api-Gateway-Timecapsule_models.UpdateCommentModel'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update Comment
      tags:
      - Comments
...
securityDefinitions:
  ApiKeyAuth:
    description: Description for what is this security definition being used
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"